bugs present
having issues loading the correct resources. the code is trying to update the detail view when an item is clicked in the master view, but the detail view is not updated. Also, the backstack code is present, but up navigation still does not work.

Something weird is going on here