a few bug fixes
1. The if/else statement in the toggle button listener was semantically flipped, so I just swapped the existing code and it worked.
2. The alarm was being set twice. You can listen to the alarm tone and notice it sounds distorted, which is because of the double playing.
I tried to fix this by commenting out the code in the Alarm.java class that sets an intent to the AlarmService, but you can still hear the alarm
going off twice. The reason we can comment that code out is that the toggle button code gets called on initialization, so that is always setting
or unsetting the alarm by itself anyways.